misc credo/dialyzer fixes
