+++ .gitignore
@@ -26,9 +26,10 @@ k8s-*.tar
 
 .iex.exs
 
-# Ignore master version as its always pulled w/ make test/all
-test/support/swagger/master.json
-
 # Ignore dialyzer files
 /priv/plts/*.plt
 /priv/plts/*.plt.hash
+
+# Ignore local k8s config/integration test files
+test/support/swagger/master.json
+integration.yaml
